requirements.txt
# Core
python-dotenv
psycopg2-binary
SQLAlchemy
redis

# Data
akshare
tushare
pandas
numpy

# Strategy & Backtesting
pandas-ta
backtrader

# Scheduling
APScheduler

# Presentation
streamlit
flask
plotly
